rpmsg: glink: Remove IRQF_NO_SUSPEND
Currently interrupts with hardware wake capabilities still cause the
CPUs to handle the interrupt at the GIC layer when IRQF_NO_SUSPEND is
used. This is reproduced in both s2idle and mem suspend.

Remove IRQF_NO_SUSPEND and add enable_irq_wake to properly account for
the wakeup interrupt.

Change-Id: Ieb8de40102fb402ca2f040020f56b8e7c42b3611
Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Ananth Raghavan Subramanian <sananth@codeaurora.org>
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>
Signed-off-by: farizmaul <ricoayuba88@gmail.com>